requirements.txt
python-telegram-bot
PySocks
Telethon-sync
requests
pandas
numpy
scikit-learn
ccxt
